requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycups==2.0.1
six==1.16.0
Unidecode==1.3.6
Werkzeug==2.3.4
zipp==3.15.0
